Nathanael sprint2 (#81)
* Changed Line Chart to Bar Chart and Updated Graphs

Changed Line Chart to Bar Chart and changed it so that it shows the total for a date

Also changed the titles for the graphs to specifically state for the current month

* Updated Transaction History on Welcome Page

Updated transaction history on welcome page so that it only shows the transactions for this month instead of all transactions - all transactions can still be seen from the transaction history page

* Fixed help.js and started email for budget/spending limits

Fixed help form on help.js and emailjs to correctly send emails with the senders name and email

Started code for emailjs to send an email once a month when a user has either spent more money than income or when exceeded budget

* Added Articles.js for Articles Page

Added base search function attached to google search - css needs to still be fixed for this page

* Updated

* Added Search Function to Savings Page

* Fixed Spending by Categories Graph on Welcome Page

* Added Articles Page

Created Articles Page and link to it from Welcome Page to move the code that I originally put in savings page. Savings page will be used for some other code.